Experimentation
Experimented with app, and adding various cornerRadius values and
CGAffineTransformations to make the presentation more interesting, but
found the frame rate dropped significantly with even a simple
cornerRadius value of 1/3 the height of the squares created. Therefore,
I will be looking to remake this app using Metal, I’d planned to use
OpenGL ES, and will still do so, but the WWDC ’14 demo of Metal really
caught my attention.